// src/gen/types/GymWorkout.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { GymWorkoutExercise } from "./GymWorkoutExercise.js";

export type GymWorkout = {
    /**
     * @type string, uuid
    */
    id: string;
    /**
     * @type string
    */
    name: string;
    /**
     * @type array
    */
    muscleGroups: string[];
    /**
     * @description Dias de treino (0=Dom … 6=Sáb)
     * @type array
    */
    daysOfWeek: number[];
    /**
     * @type array
    */
    exercises: GymWorkoutExercise[];
    /**
     * @type string, date
    */
    lastDone?: string | null;
    /**
     * @type integer
    */
    order: number;
};