// src/gen/types/PostWebsitesGymLogs.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { GymLog } from "./GymLog.js";

/**
 * @description Já existia (idempotência)
*/
export type PostWebsitesGymLogs200 = GymLog;

/**
 * @description Sessão gravada
*/
export type PostWebsitesGymLogs201 = GymLog;

export type PostWebsitesGymLogsMutationRequest = {
    /**
     * @type string, uuid
    */
    workoutId?: string | null;
    /**
     * @type string
    */
    workoutName: string;
    /**
     * @type string | undefined, date
    */
    date?: string;
    /**
     * @type integer | undefined
    */
    durationMin?: number;
    /**
     * @type integer | undefined
    */
    totalSets?: number;
    /**
     * @description UUID gerado no telemóvel para dedupe offline
     * @type string | undefined, uuid
    */
    clientUuid?: string;
    /**
     * @type array | undefined
    */
    entries?: {
        /**
         * @type string | undefined
        */
        exerciseName?: string;
        /**
         * @type string | undefined
        */
        group?: string;
        /**
         * @type array | undefined
        */
        sets?: {
            /**
             * @type number | undefined
            */
            weight?: number;
            /**
             * @type integer | undefined
            */
            reps?: number;
            /**
             * @type boolean | undefined
            */
            done?: boolean;
        }[];
    }[];
};

export type PostWebsitesGymLogsMutationResponse = (PostWebsitesGymLogs200 | PostWebsitesGymLogs201);

export type PostWebsitesGymLogsMutation = {
    Response: PostWebsitesGymLogs200 | PostWebsitesGymLogs201;
    Request: PostWebsitesGymLogsMutationRequest;
    Errors: any;
};